archive/legacy/contest_package/requirements.txt
# Minimal requirements for MangaGen contest package
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.1.0
